refactor(WarpstoneCommand): merge cmd functions into main class
Merge command functions into WarpstoneCommand class
